+++ src/bridge/SDK.ts
@@ -15,14 +15,14 @@ import {IframeBridge, IframeWrapper} from "@netless/iframe-bridge";
 import {logger, enableReport} from "../utils/Logger";
 import {convertBound} from "../utils/BoundConvert";
 import { addManagerListener } from "./Manager";
-import { RoomCallbackHandler } from "./RoomCallbackHandler";
+import { RoomCallbackHandler } from "../native/RoomCallbackHandler";
 import { addBridgeLogHook, createPageState } from "../utils/Funs";
-import { lastSchedule, ReplayerCallbackHandler } from "./ReplayerCallbackHandler";
+import { lastSchedule, ReplayerCallbackHandler } from "../native/ReplayerCallbackHandler";
 import CombinePlayerFactory from "@netless/combine-player";
-import { registerBridgeRoom } from "./RoomBridge";
-import { registerPlayerBridge } from "./PlayerBridge";
-import { Rtc } from './Rtc';
-import { SDKCallbackHandler } from './SDKCallbackHandler';
+import { registerBridgeRoom } from "./Room";
+import { registerPlayerBridge } from "./Player";
+import { Rtc } from '../Rtc';
+import { SDKCallbackHandler } from '../native/SDKCallbackHandler';
 
 let sdk: WhiteWebSdk | undefined = undefined;
 let room: Room | undefined = undefined;
