+++ src/index.tsx
@@ -2,13 +2,13 @@ import "core-js/stable";
 import "regenerator-runtime/runtime";
 import "./Polyfill";
 import "./Global";
-import "./bridge/WebSocket";
+import "./WebSocket";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import App from "./App";
 import { addDebugFunctions } from "./Debug";
 import { WindowManager } from '@netless/window-manager';
-import { registerSDKBridge } from "./bridge/SDKBridge";
+import { registerSDKBridge } from "./bridge/SDK";
 
 ReactDOM.render(
   <App />,
